+++ client/features/auth-pages/SignInPage.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { Button, Box, Typography, Link, TextField } from '@mui/material';
 
 import styles from '@styles/layout/AuthStyles.module.css';
-import { useGoogleAuthMutation } from '@/mutations/googleAuthMutation';
+import { useGoogleAuthMutation } from '@/hooks/googleAuthMutation';
 import useDocumentTitle from '@hooks/useDocumentTitle';
 
 const SignInPage = () => {
